fix route exporting page
- Make it work lol
- use convert api to handle most of it
- fix up client side routes a bit
- more separation when converting multiple geos to text
- logout button and functionality
- separate login route
- make sure to fetch config for all routes

+++ server/model/src/point_struct.rs
@@ -80,7 +80,7 @@ impl ToCollection for PointStruct {
 
 impl ToText for PointStruct {
     fn to_text(self, sep_1: &str, sep_2: &str) -> String {
-        format!("{}{}{}{}", self.lat, sep_1, self.lon, sep_2)
+        format!("{}{}{}{}", self.lat as f32, sep_1, self.lon as f32, sep_2)
     }
 }
